% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractParsedResource.R
\name{AbstractParsedResource}
\alias{AbstractParsedResource}
\title{AbstractParsedResource}
\description{
AbstractParsedResource

AbstractParsedResource
}
\section{Super classes}{
\code{\link[txronaparser:AbstractClass]{txronaparser::AbstractClass}} -> \code{\link[txronaparser:AbstractResource]{txronaparser::AbstractResource}} -> \code{AbstractParsedResource}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AbstractParsedResource$new()}}
\item \href{#method-extract}{\code{AbstractParsedResource$extract()}}
\item \href{#method-clone}{\code{AbstractParsedResource$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="etl">}\href{../../txronaparser/html/AbstractResource.html#method-etl}{\code{txronaparser::AbstractResource$etl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="load">}\href{../../txronaparser/html/AbstractResource.html#method-load}{\code{txronaparser::AbstractResource$load()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="spawn">}\href{../../txronaparser/html/AbstractResource.html#method-spawn}{\code{txronaparser::AbstractResource$spawn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="transform">}\href{../../txronaparser/html/AbstractResource.html#method-transform}{\code{txronaparser::AbstractResource$transform()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractParsedResource$new(raw_data = NULL, can_spawn = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw_data}}{raw_data}

\item{\code{can_spawn}}{can_spawn}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract"></a>}}
\if{latex}{\out{\hypertarget{method-extract}{}}}
\subsection{Method \code{extract()}}{
extract
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractParsedResource$extract(force = private$.should_extract)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force}}{force}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractParsedResource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
