% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractResource.R
\name{AbstractResource}
\alias{AbstractResource}
\title{AbstractResource}
\description{
AbstractResource

AbstractResource
}
\section{Super class}{
\code{\link[txronaparser:AbstractClass]{txronaparser::AbstractClass}} -> \code{AbstractResource}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{}

\item{\code{value}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-extract}{\code{AbstractResource$extract()}}
\item \href{#method-transform}{\code{AbstractResource$transform()}}
\item \href{#method-load}{\code{AbstractResource$load()}}
\item \href{#method-etl}{\code{AbstractResource$etl()}}
\item \href{#method-spawn}{\code{AbstractResource$spawn()}}
\item \href{#method-clone}{\code{AbstractResource$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractClass" data-id="initialize">}\href{../../txronaparser/html/AbstractClass.html#method-initialize}{\code{txronaparser::AbstractClass$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract"></a>}}
\if{latex}{\out{\hypertarget{method-extract}{}}}
\subsection{Method \code{extract()}}{
extract
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractResource$extract()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
transform
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractResource$transform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
load
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractResource$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-etl"></a>}}
\if{latex}{\out{\hypertarget{method-etl}{}}}
\subsection{Method \code{etl()}}{
etl
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractResource$etl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-spawn"></a>}}
\if{latex}{\out{\hypertarget{method-spawn}{}}}
\subsection{Method \code{spawn()}}{
spawn
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractResource$spawn(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{spread}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AbstractResource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
