% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{bootstrap}
\alias{bootstrap}
\title{Attempts to load config and sets up the working directory.}
\usage{
bootstrap(config_file_path = NULL)
}
\arguments{
\item{config_file_path}{The input sent to load_config}
}
\description{
Attempts to load config and sets up the working directory.
}
\seealso{
\code{\link{load_config}}
}
