% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CONFIG_SCHEMA_PATH}
\alias{CONFIG_SCHEMA_PATH}
\title{Where the config schema can be found}
\format{
An object of class \code{character} of length 1.
}
\usage{
CONFIG_SCHEMA_PATH
}
\description{
Where the config schema can be found
}
\keyword{datasets}
