% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DATA_LIST_TYPES}
\alias{DATA_LIST_TYPES}
\title{All the available data list types}
\format{
An object of class \code{character} of length 3.
}
\usage{
DATA_LIST_TYPES
}
\description{
All the available data list types
}
\keyword{datasets}
