% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{default_config}
\alias{default_config}
\title{Load the default config shipped with the package}
\usage{
default_config()
}
\value{
the default config
}
\description{
Load the default config shipped with the package
}
