% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DEFAULT_CONFIG_FILE_PATH}
\alias{DEFAULT_CONFIG_FILE_PATH}
\title{Where the default config file can be found}
\format{
An object of class \code{character} of length 1.
}
\usage{
DEFAULT_CONFIG_FILE_PATH
}
\description{
Where the default config file can be found
}
\keyword{datasets}
