% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\docType{data}
\name{ERROR_IS_ABSTRACT_METHOD}
\alias{ERROR_IS_ABSTRACT_METHOD}
\title{ERROR_IS_ABSTRACT_METHOD}
\format{
An object of class \code{character} of length 1.
}
\usage{
ERROR_IS_ABSTRACT_METHOD
}
\description{
ERROR_IS_ABSTRACT_METHOD
}
\keyword{datasets}
