% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_parser_county_cases.R
\name{etl_parser_county_cases}
\alias{etl_parser_county_cases}
\title{ETL parser for county case data}
\usage{
etl_parser_county_cases(raw_spreadsheet_data)
}
\arguments{
\item{raw_spreadsheet_data}{The raw data to parse}
}
\value{
\itemize{
  \item \code{population_data} (county name, population)
  \item \code{total_case_data} (data, cases in county a1, cases in county a2, ...)
}
}
\description{
ETL parser for county case data
}
