% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{generate_new_spreadsheets_config}
\alias{generate_new_spreadsheets_config}
\title{Ensures all discovered spreadsheets are in the config file}
\usage{
generate_new_spreadsheets_config(spreadsheet_urls = parse_spreadsheet_urls())
}
\description{
Ensures all discovered spreadsheets are in the config file
}
