% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_requests.R
\name{get_etl_parser}
\alias{get_etl_parser}
\title{Either loads an existing ETL parser or returns an anonymous function that does nothing}
\usage{
get_etl_parser(etl_parser_name)
}
\description{
Either loads an existing ETL parser or returns an anonymous function that does nothing
}
