% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-data-io.R
\name{load_all_data_local}
\alias{load_all_data_local}
\title{Loads all data on disk into the singleton \code{.GlobalEnv$all_data}}
\usage{
load_all_data_local()
}
\description{
Loads all data on disk into the singleton \code{.GlobalEnv$all_data}
}
