% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_requests.R
\name{load_all_spreadsheet_data}
\alias{load_all_spreadsheet_data}
\title{Loads and parses all the remote spreadsheets into the singleton \code{.GlobalEnv$all_data}}
\usage{
load_all_spreadsheet_data(spreadsheet_config_list = trp_config$spreadsheets)
}
\arguments{
\item{spreadsheet_config_list}{A list of spreadsheet configs as seen in the config file}
}
\description{
Loads and parses all the remote spreadsheets into the singleton \code{.GlobalEnv$all_data}
}
