% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Loads a config file. If nothing is passed, it loads the default. If a
directory is passed, it looks for `config.yml` in that directory. If anything
else is passed, it will attempt to load that as a file. If any of that fails,
this will return the defaults.}
\usage{
load_config(config_file_path = NULL)
}
\arguments{
\item{config_file_path}{A path}
}
\description{
The result is loaded as `.GlobalEnv$trp_config`.
}
