% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-data-io.R
\name{load_data_list_local}
\alias{load_data_list_local}
\title{Loads an entire data list type}
\usage{
load_data_list_local(data_list_type = "misc")
}
\arguments{
\item{data_list_type}{The data list type to load}
}
\value{
The entire loaded list
}
\description{
Loads an entire data list type
}
\seealso{
\code{\link{DATA_LIST_TYPES}}
}
