% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_requests.R
\name{load_remote_spreadsheet}
\alias{load_remote_spreadsheet}
\title{Loads a remote spreadsheet using the insecure handler}
\usage{
load_remote_spreadsheet(spreadsheet_url, handle = insecure_handler_factory())
}
\description{
Loads a remote spreadsheet using the insecure handler
}
\note{
May need to include ways to pass args to readxl
}
