% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-data-io.R
\name{load_single_data_local}
\alias{load_single_data_local}
\title{Loads a single data object from either its name and list type or a full path}
\usage{
load_single_data_local(
  data_name,
  data_list_type = "misc",
  data_file_path = NULL
)
}
\arguments{
\item{data_name}{The object name}

\item{data_list_type}{The data list type to load}

\item{data_full_path}{Default \code{NULL}; if set will be used as the RDS path}
}
\value{
The contents of the loaded RDS file
}
\description{
Loads a single data object from either its name and list type or a full path
}
\seealso{
\code{\link{DATA_LIST_TYPES}}
}
