% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_requests.R
\name{load_single_spreadsheet_data}
\alias{load_single_spreadsheet_data}
\title{Load and parse a single remote spreadsheet}
\usage{
load_single_spreadsheet_data(spreadsheet_name)
}
\arguments{
\item{spreadsheet_name}{The name (from the config list) of the spreadsheet to load}
}
\value{
\itemize{
  \item \code{raw} The unparsed spreadsheet
  \item \code{parsed} If this spreadsheet has an ETL parser defined, then its return value. Otherwise empty.
}
}
\description{
Load and parse a single remote spreadsheet
}
