% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{nuke_content}
\alias{nuke_content}
\title{Completely destroys the working directory}
\usage{
nuke_content(are_you_sure = FALSE)
}
\arguments{
\item{are_you_sure}{Requires some thought before destruction}
}
\description{
Completely destroys the working directory
}
