% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_requests.R
\name{parse_spreadsheet_urls}
\alias{parse_spreadsheet_urls}
\title{Finds all the Texas spreadsheets on the page}
\usage{
parse_spreadsheet_urls(data_html = scrape_source_page())
}
\description{
Finds all the Texas spreadsheets on the page
}
