% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz_ParsedResource_CasesoverTimebyCounty.R
\name{ParsedResourceCasesoverTimebyCounty}
\alias{ParsedResourceCasesoverTimebyCounty}
\title{ParsedResourceCasesoverTimebyCounty}
\description{
ParsedResourceCasesoverTimebyCounty

ParsedResourceCasesoverTimebyCounty
}
\section{Super classes}{
\code{\link[txronaparser:AbstractClass]{txronaparser::AbstractClass}} -> \code{\link[txronaparser:AbstractResource]{txronaparser::AbstractResource}} -> \code{\link[txronaparser:AbstractParsedResource]{txronaparser::AbstractParsedResource}} -> \code{ParsedResourceCasesoverTimebyCounty}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ParsedResourceCasesoverTimebyCounty$new()}}
\item \href{#method-load}{\code{ParsedResourceCasesoverTimebyCounty$load()}}
\item \href{#method-transform}{\code{ParsedResourceCasesoverTimebyCounty$transform()}}
\item \href{#method-clone}{\code{ParsedResourceCasesoverTimebyCounty$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="etl">}\href{../../txronaparser/html/AbstractResource.html#method-etl}{\code{txronaparser::AbstractResource$etl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="spawn">}\href{../../txronaparser/html/AbstractResource.html#method-spawn}{\code{txronaparser::AbstractResource$spawn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractParsedResource" data-id="extract">}\href{../../txronaparser/html/AbstractParsedResource.html#method-extract}{\code{txronaparser::AbstractParsedResource$extract()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParsedResourceCasesoverTimebyCounty$new(
  raw_data = NULL,
  can_spawn = FALSE,
  init_data = PARSED_RESOURCE_DATA[["CasesoverTimebyCounty"]]
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{raw_data}}{raw_data}

\item{\code{can_spawn}}{can_spawn}

\item{\code{init_data}}{init_data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
load
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParsedResourceCasesoverTimebyCounty$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
transform
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParsedResourceCasesoverTimebyCounty$transform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ParsedResourceCasesoverTimebyCounty$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
