% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/zzzz_RawResource_EstimatedActiveCasesoverTimebyCounty.R
\name{RawResourceEstimatedActiveCasesoverTimebyCounty}
\alias{RawResourceEstimatedActiveCasesoverTimebyCounty}
\title{RawResourceEstimatedActiveCasesoverTimebyCounty}
\description{
RawResourceEstimatedActiveCasesoverTimebyCounty

RawResourceEstimatedActiveCasesoverTimebyCounty
}
\section{Super classes}{
\code{\link[txronaparser:AbstractClass]{txronaparser::AbstractClass}} -> \code{\link[txronaparser:AbstractResource]{txronaparser::AbstractResource}} -> \code{\link[txronaparser:AbstractRawResource]{txronaparser::AbstractRawResource}} -> \code{RawResourceEstimatedActiveCasesoverTimebyCounty}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-transform}{\code{RawResourceEstimatedActiveCasesoverTimebyCounty$transform()}}
\item \href{#method-load}{\code{RawResourceEstimatedActiveCasesoverTimebyCounty$load()}}
\item \href{#method-clone}{\code{RawResourceEstimatedActiveCasesoverTimebyCounty$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="etl">}\href{../../txronaparser/html/AbstractResource.html#method-etl}{\code{txronaparser::AbstractResource$etl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="spawn">}\href{../../txronaparser/html/AbstractResource.html#method-spawn}{\code{txronaparser::AbstractResource$spawn()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractRawResource" data-id="extract">}\href{../../txronaparser/html/AbstractRawResource.html#method-extract}{\code{txronaparser::AbstractRawResource$extract()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractRawResource" data-id="initialize">}\href{../../txronaparser/html/AbstractRawResource.html#method-initialize}{\code{txronaparser::AbstractRawResource$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
transform
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawResourceEstimatedActiveCasesoverTimebyCounty$transform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
load
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawResourceEstimatedActiveCasesoverTimebyCounty$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RawResourceEstimatedActiveCasesoverTimebyCounty$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
