% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{reload_data_remote}
\alias{reload_data_remote}
\title{Convenience method to refresh everything from the remote}
\usage{
reload_data_remote()
}
\description{
Convenience method to refresh everything from the remote
}
