% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceOrchestrator.R
\name{ResourceOrchestrator}
\alias{ResourceOrchestrator}
\title{ResourceOrchestrator}
\description{
ResourceOrchestrator

ResourceOrchestrator
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-e2e_automated}{\code{ResourceOrchestrator$e2e_automated()}}
\item \href{#method-clone}{\code{ResourceOrchestrator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-e2e_automated"></a>}}
\if{latex}{\out{\hypertarget{method-e2e_automated}{}}}
\subsection{Method \code{e2e_automated()}}{
e2e_automated
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceOrchestrator$e2e_automated()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResourceOrchestrator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
