% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{data}
\name{RONA_WORLD}
\alias{RONA_WORLD}
\title{The finished environment}
\format{
An object of class \code{environment} of length 0.
}
\usage{
RONA_WORLD
}
\description{
The finished environment
}
\keyword{datasets}
