% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RootResource.R
\name{RootResource}
\alias{RootResource}
\title{RootResource}
\description{
RootResource

RootResource
}
\section{Super classes}{
\code{\link[txronaparser:AbstractClass]{txronaparser::AbstractClass}} -> \code{\link[txronaparser:AbstractResource]{txronaparser::AbstractResource}} -> \code{RootResource}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RootResource$new()}}
\item \href{#method-extract}{\code{RootResource$extract()}}
\item \href{#method-transform}{\code{RootResource$transform()}}
\item \href{#method-load}{\code{RootResource$load()}}
\item \href{#method-clone}{\code{RootResource$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="etl">}\href{../../txronaparser/html/AbstractResource.html#method-etl}{\code{txronaparser::AbstractResource$etl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="txronaparser" data-topic="AbstractResource" data-id="spawn">}\href{../../txronaparser/html/AbstractResource.html#method-spawn}{\code{txronaparser::AbstractResource$spawn()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootResource$new(can_spawn = FALSE, init_data = ROOT_RESOURCE_DATA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{can_spawn}}{can_spawn}

\item{\code{init_data}}{init_data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract"></a>}}
\if{latex}{\out{\hypertarget{method-extract}{}}}
\subsection{Method \code{extract()}}{
extract
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootResource$extract()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-transform"></a>}}
\if{latex}{\out{\hypertarget{method-transform}{}}}
\subsection{Method \code{transform()}}{
transform
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootResource$transform()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-load"></a>}}
\if{latex}{\out{\hypertarget{method-load}{}}}
\subsection{Method \code{load()}}{
load
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootResource$load()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RootResource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
