% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-data-io.R
\name{save_all_data_local}
\alias{save_all_data_local}
\title{Saves all data present in the singleton \code{all_data}}
\usage{
save_all_data_local(new_data = all_data)
}
\arguments{
\item{new_data}{The data to store; default is \code{all_data}}
}
\description{
Saves all data present in the singleton \code{all_data}
}
