% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-data-io.R
\name{save_data_list_local}
\alias{save_data_list_local}
\title{Saves a specific data list}
\usage{
save_data_list_local(data_list, data_list_type = "misc")
}
\arguments{
\item{data_list}{The list to walk over and store}

\item{data_list_type}{The data list type to store it with}
}
\description{
Saves a specific data list
}
\seealso{
\code{\link{DATA_LIST_TYPES}}
}
