% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local-data-io.R
\name{save_single_data_local}
\alias{save_single_data_local}
\title{Saves a single data object}
\usage{
save_single_data_local(data_name, data_contents, data_list_type = "misc")
}
\arguments{
\item{data_name}{The object name}

\item{data_contents}{The actual value}

\item{data_list_type}{The data list type to store it with}
}
\description{
Saves a single data object
}
\seealso{
\code{\link{DATA_LIST_TYPES}}
}
