% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_requests.R
\name{scrape_source_page}
\alias{scrape_source_page}
\title{Scrapes the DSHS page for spreadsheets and returns the HTML}
\usage{
scrape_source_page(
  url_to_scrape = trp_config$dshs_data_source$url,
  handle = insecure_handler_factory()
)
}
\description{
Scrapes the DSHS page for spreadsheets and returns the HTML
}
