% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{validate_config}
\alias{validate_config}
\title{Validates a given config against the config schema. It will silently succeed
or loudly fail.}
\usage{
validate_config(config_to_validate)
}
\arguments{
\item{config_to_validate}{Some config object to validate}
}
\description{
Validates a given config against the config schema. It will silently succeed
or loudly fail.
}
