% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\docType{data}
\name{WARNING_NUKE_CONTENT}
\alias{WARNING_NUKE_CONTENT}
\title{Warning for nuke_content without \code{are_you_sure}}
\format{
An object of class \code{character} of length 1.
}
\usage{
WARNING_NUKE_CONTENT
}
\description{
Warning for nuke_content without \code{are_you_sure}
}
\seealso{
\code{\link{nuke_content}}
}
\keyword{datasets}
