% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_config}
\alias{write_config}
\title{Writes the provided config object to the proper location}
\usage{
write_config(config = NULL)
}
\arguments{
\item{config}{A config object}
}
\description{
Writes the provided config object to the proper location
}
