% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\docType{data}
\name{XPATH_SPREADSHEET_URLS}
\alias{XPATH_SPREADSHEET_URLS}
\title{XPATH_SPREADSHEET_URLS}
\format{
An object of class \code{character} of length 1.
}
\usage{
XPATH_SPREADSHEET_URLS
}
\description{
XPATH_SPREADSHEET_URLS
}
\keyword{datasets}
